requirements.txt
telethon
PyYAML
